requirements.txt
fastapi 
uvicorn 
sqlalchemy 
passlib[bcrypt] 
python-jose[cryptography] 
python-dotenv 
python-multipart 
sqlalchemy_utils 
alembic 
bcrypt==4.3.0
pydantic[email]

langchain_ollama
youtube_transcript_api
langchain